+++ pltraining-puppetfactory/ReleaseNotes.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Added 'redeploy' command to redeploy environment
+
 ## Version 0.4.9
 
 * Narrowed race condition in spec tests